Cleaned up redundancy in runAllStuff()
There was a lot of unnecessary duplication of code in runAllStuff()
The point of this funciton is to use the other functions in the
class, in a way that can be generic and readable.
So I refactored the whole function by making it use class functions.
This also eliminated the need for fresh validation generators,
I think that "bug" may have resulted from having to many versions of
your variables.