Added a modelRunner and implemented model1 to abstract version.
Added a modelRunner script that can be used to run the models.
It calls the runAllStuff for each model, and I think will be useful
for testing.
Then debugged the model1.py and made a few changes to model2.py
so that they play nicely with the common abstract version of
common methods.
trainModel is also moved to the abstract class, although there is
a version in model1.py still to debug why the quantized models for
model 1 are giving random guessing.